+++ Temp.java
@@ -4,7 +4,7 @@ class Temp{
     public static void main(String args[]){
         Scanner scanner = new Scanner(System.in);
         System.out.print("Enter the temperature in celsius: ");
-        float n1 = scanner.nextInt();
+        float n1 = scanner.nextFloat();
         System.out.println("The temperature in fahrenheit is = "+(9 * n1 + 32));
         scanner.close();
     }
